element comments and types
